friends_network/requirements.txt
asgiref==3.7.2
attrs==23.2.0
cffi==1.16.0
cryptography==42.0.1
Django==5.0.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.17.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
